+++ packages/plugin-i18n/CHANGELOG.md
@@ -1,5 +1,11 @@
 # build-plugin-ice-i18n
 
+## 0.1.1
+
+- fix: fail to redirect to another route in CSR
+- fix: fail to render NoMatch component which has no specified path
+- chore: rename `redirect` to `autoRedirect`
+
 ## 0.1.0
 
 - feat: init plugin